Create dps5020_Test.ino
Basic code to send commands to a Rui Deng programmable power supply . Implements a class to represent a supply. This example code will control a device connected to the hardware serial port (Pins 0 and 1 and the Serial1 object in the Arduino IDE) of an ATmega32U4 based microcontroller board like the Pololu A-Star series. This code only sends commands to the supply, it does not handle any responses returned.