Update RCTech_PS_Comm.cpp
Changed the DPS_powerSupply::sendData and DPS_powerSupply::sendData2 methods to use the _stream object. They were hardcoded to use Serial1 before this. I think that when I was writing this originally, I couldn't figure out the correct syntax to call the write() method of the pointer to the Stream object in the class. 
This seems to work, though:
_stream->write(message, 8);